Added sales_count property
